// src/shared/api/models/patchApiBundleOffersIdStatusBodyStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

/**
 * New status for the bundle offer
 */
export type PatchApiBundleOffersIdStatusBodyStatus =
    (typeof PatchApiBundleOffersIdStatusBodyStatus)[keyof typeof PatchApiBundleOffersIdStatusBodyStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PatchApiBundleOffersIdStatusBodyStatus = {
    BLOCKED: 'BLOCKED',
    EXPIRED: 'EXPIRED',
    USED: 'USED',
    UNUSED: 'UNUSED',
    PARTIALLY_USED: 'PARTIALLY_USED',
} as const;